backend/requirements.txt
fastapi
uvicorn
sqlalchemy
requests
python-dotenv
psycopg2-binary
